Project structure changes for better maintanance
