+++ components/food/DiseaseItem.js
