+++ pages/disease.js
@@ -1,11 +1,11 @@
 import { withRouter } from "next/router";
 import Data from "../components/dataList";
-import Layout from "../components/Layout";
+import Layout from "../components/layout/Layout";
 import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
-import FoodList from "../components/FoodList";
-import Image from "../components/ContainerImage";
-import Breadcrumb from "../components/Breadcrumb";
+import FoodList from "../components/food/FoodList";
+import Image from "../components/disease/ContainerImage";
+import Breadcrumb from "../components/layout/Breadcrumb";
 import Link from "next/link";
 import Button from "@material-ui/core/Button";
